rustc_metadata: Move native library search code to `rustc_codegen_ssa`
