+++ compiler/rustc_metadata/src/native_libs.rs
@@ -1,6 +1,3 @@
-use std::ops::ControlFlow;
-use std::path::{Path, PathBuf};
-
 use rustc_abi::ExternAbi;
 use rustc_attr_parsing::eval_config_entry;
 use rustc_data_structures::fx::FxHashSet;
@@ -15,157 +12,11 @@ use rustc_session::Session;
 use rustc_session::cstore::{
     DllCallingConvention, DllImport, DllImportSymbolType, ForeignModule, NativeLib,
 };
-use rustc_session::search_paths::PathKind;
 use rustc_span::Symbol;
 use rustc_span::def_id::{DefId, LOCAL_CRATE};
-use rustc_target::spec::{Arch, BinaryFormat, CfgAbi, Env, LinkSelfContainedComponents, Os};
+use rustc_target::spec::{Arch, BinaryFormat, CfgAbi};
 
 use crate::diagnostics;
-
-/// The fallback directories are passed to linker, but not used when rustc does the search,
-/// because in the latter case the set of fallback directories cannot always be determined
-/// consistently at the moment.
-pub struct NativeLibSearchFallback<'a> {
-    pub self_contained_components: LinkSelfContainedComponents,
-    pub apple_sdk_root: Option<&'a Path>,
-}
-
-pub fn walk_native_lib_search_dirs<R>(
-    sess: &Session,
-    fallback: Option<NativeLibSearchFallback<'_>>,
-    mut f: impl FnMut(&Path, bool /*is_framework*/) -> ControlFlow<R>,
-) -> ControlFlow<R> {
-    // Library search paths explicitly supplied by user (`-L` on the command line).
-    for search_path in sess.target_filesearch().cli_search_paths(PathKind::Native) {
-        f(&search_path.dir, false)?;
-    }
-    for search_path in sess.target_filesearch().cli_search_paths(PathKind::Framework) {
-        // Frameworks are looked up strictly in framework-specific paths.
-        if search_path.kind != PathKind::All {
-            f(&search_path.dir, true)?;
-        }
-    }
-
-    let Some(NativeLibSearchFallback { self_contained_components, apple_sdk_root }) = fallback
-    else {
-        return ControlFlow::Continue(());
-    };
-
-    // The toolchain ships some native library components and self-contained linking was enabled.
-    // Add the self-contained library directory to search paths.
-    if self_contained_components.intersects(
-        LinkSelfContainedComponents::LIBC
-            | LinkSelfContainedComponents::UNWIND
-            | LinkSelfContainedComponents::MINGW,
-    ) {
-        f(&sess.target_tlib_path.dir.join("self-contained"), false)?;
-    }
-
-    let has_shared_llvm_apple_darwin =
-        sess.target.is_like_darwin && sess.target_tlib_path.dir.join("libLLVM.dylib").exists();
-
-    // Toolchains for some targets may ship `libunwind.a`, but place it into the main sysroot
-    // library directory instead of the self-contained directories.
-    // Sanitizer libraries have the same issue and are also linked by name on Apple targets.
-    // The targets here should be in sync with `copy_third_party_objects` in bootstrap.
-    // On Apple targets, shared LLVM is linked by name, so when `libLLVM.dylib` is
-    // present in the target libdir, add that directory to the linker search path.
-    // FIXME: implement `-Clink-self-contained=+/-unwind,+/-sanitizers`, move the shipped libunwind
-    // and sanitizers to self-contained directory, and stop adding this search path.
-    // FIXME: On AIX this also has the side-effect of making the list of library search paths
-    // non-empty, which is needed or the linker may decide to record the LIBPATH env, if
-    // defined, as the search path instead of appending the default search paths.
-    if sess.target.cfg_abi == CfgAbi::Fortanix
-        || sess.target.os == Os::Linux
-        || sess.target.os == Os::Fuchsia
-        || sess.target.is_like_aix
-        || sess.target.is_like_darwin
-            && (!sess.sanitizers().is_empty() || has_shared_llvm_apple_darwin)
-        || sess.target.os == Os::Windows
-            && sess.target.env == Env::Gnu
-            && sess.target.cfg_abi == CfgAbi::Llvm
-    {
-        f(&sess.target_tlib_path.dir, false)?;
-    }
-
-    // Mac Catalyst uses the macOS SDK, but to link to iOS-specific frameworks
-    // we must have the support library stubs in the library search path (#121430).
-    if let Some(sdk_root) = apple_sdk_root
-        && sess.target.env == Env::MacAbi
-    {
-        f(&sdk_root.join("System/iOSSupport/usr/lib"), false)?;
-        f(&sdk_root.join("System/iOSSupport/System/Library/Frameworks"), true)?;
-    }
-
-    ControlFlow::Continue(())
-}
-
-pub fn try_find_native_static_library(
-    sess: &Session,
-    name: &str,
-    verbatim: bool,
-) -> Option<PathBuf> {
-    let default = sess.staticlib_components(verbatim);
-    let formats = if verbatim {
-        vec![default]
-    } else {
-        // On Windows, static libraries sometimes show up as libfoo.a and other
-        // times show up as foo.lib
-        let unix = ("lib", ".a");
-        if default == unix { vec![default] } else { vec![default, unix] }
-    };
-
-    walk_native_lib_search_dirs(sess, None, |dir, is_framework| {
-        if !is_framework {
-            for (prefix, suffix) in &formats {
-                let test = dir.join(format!("{prefix}{name}{suffix}"));
-                if test.exists() {
-                    return ControlFlow::Break(test);
-                }
-            }
-        }
-        ControlFlow::Continue(())
-    })
-    .break_value()
-}
-
-pub fn try_find_native_dynamic_library(
-    sess: &Session,
-    name: &str,
-    verbatim: bool,
-) -> Option<PathBuf> {
-    let default = sess.staticlib_components(verbatim);
-    let formats = if verbatim {
-        vec![default]
-    } else {
-        // While the official naming convention for MSVC import libraries
-        // is foo.lib, Meson follows the libfoo.dll.a convention to
-        // disambiguate .a for static libraries
-        let meson = ("lib", ".dll.a");
-        // and MinGW uses .a altogether
-        let mingw = ("lib", ".a");
-        vec![default, meson, mingw]
-    };
-
-    walk_native_lib_search_dirs(sess, None, |dir, is_framework| {
-        if !is_framework {
-            for (prefix, suffix) in &formats {
-                let test = dir.join(format!("{prefix}{name}{suffix}"));
-                if test.exists() {
-                    return ControlFlow::Break(test);
-                }
-            }
-        }
-        ControlFlow::Continue(())
-    })
-    .break_value()
-}
-
-pub fn find_native_static_library(name: &str, verbatim: bool, sess: &Session) -> PathBuf {
-    try_find_native_static_library(sess, name, verbatim).unwrap_or_else(|| {
-        sess.dcx().emit_fatal(diagnostics::MissingNativeLibrary::new(name, verbatim))
-    })
-}
 
 pub(crate) fn collect(tcx: TyCtxt<'_>, LocalCrate: LocalCrate) -> Vec<NativeLib> {
     let mut collector = Collector { tcx, libs: Vec::new() };
